+++ src/harlequin/exception.py
@@ -33,6 +33,10 @@ class HarlequinWizardError(HarlequinError):
     pass
 
 
+class HarlequinTzDataError(HarlequinError):
+    pass
+
+
 def pretty_print_error(error: HarlequinError) -> None:
     from rich import print
     from rich.panel import Panel